feature: add first solution about product

+++ src/main/java/com/backend/d2/dtos/products/responses/ProductCreateResponseDTO.java
@@ -1,21 +1,20 @@
 package com.backend.d2.dtos.products.responses;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductCreateResponseDTO {
-
     private Long id;
     private Long categoryId;
     private Long supplierId;
     private String name;
     private String barcode;
-    private double price;
+    private BigDecimal price;
     private int stock;
     private String description;
 }
