+++ src/main/java/com/backend/d2/dtos/products/responses/ProductSearchResponseDTO.java
@@ -3,6 +3,7 @@ package com.backend.d2.dtos.products.responses;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -11,7 +12,8 @@ public class ProductSearchResponseDTO {
     private Long id;
     private String name;
     private String barcode;
-    private double price;
+    private BigDecimal price;
+
     private String category;
     private String supplier;
 }
